Add ability to show/hide boosts from accounts
Fixes some stuff in #6
