+++ routes/_api/showReblogs.js
@@ -1,0 +1,7 @@
+import { auth, basename } from './utils'
+import { post, WRITE_TIMEOUT } from '../_utils/ajax'
+
+export function setShowReblogs (instanceName, accessToken, accountId, showReblogs) {
+  let url = `${basename(instanceName)}/api/v1/accounts/${accountId}/follow`
+  return post(url, { reblogs: !!showReblogs }, auth(accessToken), {timeout: WRITE_TIMEOUT})
+}
