+++ routes/_api/user.js
@@ -1,4 +1,4 @@
-import { get, paramsString, DEFAULT_TIMEOUT } from '../_utils/ajax'
+import { get, DEFAULT_TIMEOUT } from '../_utils/ajax'
 import { auth, basename } from './utils'
 
 export function getVerifyCredentials (instanceName, accessToken) {
@@ -9,11 +9,4 @@ export function getVerifyCredentials (instanceName, accessToken) {
 export function getAccount (instanceName, accessToken, accountId) {
   let url = `${basename(instanceName)}/api/v1/accounts/${accountId}`
   return get(url, auth(accessToken), {timeout: DEFAULT_TIMEOUT})
-}
-
-export async function getRelationship (instanceName, accessToken, accountId) {
-  let url = `${basename(instanceName)}/api/v1/accounts/relationships`
-  url += '?' + paramsString({id: accountId})
-  let res = await get(url, auth(accessToken), {timeout: DEFAULT_TIMEOUT})
-  return res[0]
 }
